fix: update DingTalk user model department ID list data type
- Change deptIdList field type from List<Integer> to List<Long> in DingtalkUserModel.User
- Improve data type compatibility for DingTalk department ID handling
- Prevent potential data overflow issues with large department IDs
- Align with DingTalk API specification for department ID format

+++ cratos-eds-dingtalk/src/main/java/com/baiyi/cratos/eds/dingtalk/model/DingtalkUserModel.java
@@ -68,7 +68,7 @@ public class DingtalkUserModel {
         private String orgEmail;
         private String name;
         @JsonProperty("dept_id_list")
-        private List<Integer> deptIdList;
+        private List<Long> deptIdList;
         @JsonProperty("job_number")
         private String jobNumber;
         @JsonProperty("state_code")
